+++ CHANGELOG.rst
@@ -23,6 +23,8 @@ Many backward incompatible changes in this update, look carefully at the changel
    to ``Statechart.add_transition``.
  - (Changed) Rewrite ``io.import_from_yaml`` to load states before transitions. This function does not anymore
    accept a ``validate_statechart`` parameter. Its ``validate_schema`` parameter is now ``not ignore_schema``.
+ - (Changed) Split ``model.HistoryState`` into a mixin ``model.HistoryStateMixin`` and two concrete subclasses,
+   namely ``ShallowHistoryState`` and ``DeepHistoryState``.
  - (Removed) Contracts and codes for statecharts (define them on root state instead).
  - (Removed) Cache for several ``model.Statechart`` methods.
  - (Removed) ``Statechart.validate`` (C1, C2 and C4 are now included in the construction process).